microservicios/log_service/requirements.txt
flask
pymongo
pulsar-client